Ensure abiword package is always present
Change I9831526581a5fc1f2014c9985637664c79b937bf
moved the required packages list statement into a
conditional block, making it to be ensured only
in the case nodejs_version was set to 'system'.

This was causing the abiword library to not be
installed and then etherpad.o.o was appending the
following error message to new pads:

  Error: Abiword does not exist at this path,
  check your settings file -- To suppress these
  warning messages change suppressErrorsInPadText
  to true in your settings.json

This patch makes the abiword library to always be
installed by moving it out of that conditional
statement.

Change-Id: I50f55e4408d850b2bf449b121b9014d511518481