Fix up log rotation
The path to the logs is wrong and they're not being rotated.

I think the rotation might have ended up in site.pp as it was relying
on the service defined there.  This doesn't really seem necessary so
I've moved it back with the main configuration, and fixed the paths.

Change-Id: I99b71481e222e6f83b6be72261fa14b7a10dc62c