Turn off binary symlink in nodejs module
On Xenial, the nodejs puppet module tries to create a symlink from
/usr/bin/node to /usr/bin/nodejs. Depending on the resource ordering,
this might happen before or after the nodejs package creates its own
/usr/bin/node. If puppet has the chance to make the symlink itself, it
breaks node by creating a broken link cycle. This patch uses the
legacy_debian_symlinks parameter to prevent the puppet module from doing
that. This problem was fixed and the option was deprecated and removed
in later versions of the nodejs puppet module, so we will have to remove
the parameter if we update the nodejs module.

In this case, on puppet 3 a quirk of the resource ordering was causing
the module not to be idempotent since the second puppet run would change
the file to a symlink, so this patch has the side benefit of also
allowing us to add a test for idempotency. With that second puppet run,
the etherpad service now has enough time to fully start up and we can
uncomment the curl test.

Change-Id: Ic834ad6f0828cc2d20810e48bb053f9b14659c4a