Fix env for plugin installs
In an earlier change we set up npm installations to happen in the eplite
user's homedir rather than in the log dir. We did this because with
systemd we no longer have a log dir on all installations. The etherpad
lite plugin installation was still attemping to use the log dir which
doesn't work on newer systems.

Update this plugin installation process to use the homedir like the
normal installation process.

Also add plugin installation to the tests we run to test this works now.

Change-Id: I5cff75c89b16d61c4902c920259b82463306833e