Actually use connection-tuning configuration
The connection tuning configuration setup relied on a symlink that was
broken. Fix the symlink by fixing the target filename. Also we update
the symlink name to have a .conf suffix since everything else has that
suffix on Xenial.

Note this removes support for precise connection tuning to simplify the
problem as precise is no longer supported by ubuntu.

Change-Id: I094a43b2395aeef17175cea6db5dc2469b67ed6d